+++ packages/frontend/src/design-system/components/common/index.ts
@@ -9,3 +9,4 @@ export { default as Footer } from "./Footer";
 export { default as Info } from "./Info";
 export { CodeBlock } from "./CodeBlock";
 export { default as LinkButton } from "./LinkButton";
+export { Coachmark, type CoachmarkProps } from "./Coachmark";
